% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{elapsed_time}
\alias{elapsed_time}
\title{Returns a string with the formatted elapsed time.}
\usage{
elapsed_time(start_time, end_time)
}
\description{
Returns a string with the formatted elapsed time.
}
\keyword{internal}
