% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{iri_replacement_function}
\alias{iri_replacement_function}
\title{Create a function to replace IRI styles in strings.}
\usage{
iri_replacement_function(iri_style, prefixes, replace_in_literal)
}
\arguments{
\item{iri_style}{One of "long", "short", "mdlink", or "html".}

\item{prefixes}{Tibble with "short" and "long" prefix columns.}

\item{replace_in_literal}{If TRUE, then replacement is also done within
literal strings, not just in IRIs.}
}
\value{
A function that replaces IRIs according to the specified style.
}
\description{
Returns a function to replace IRI styles in strings.
}
\keyword{internal}
