% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{is_valid_rdf_term}
\alias{is_valid_rdf_term}
\title{Checks whether the \code{x} list object has the expected structure of a standard
SPARQL query response in JSON format.}
\usage{
is_valid_rdf_term(x)
}
\description{
Checks whether the \code{x} list object has the expected structure of a standard
SPARQL query response in JSON format.
}
\keyword{internal}
