% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{modify_iri_style}
\alias{modify_iri_style}
\title{Change the style of IRIs from "long" or "short" into another form.}
\usage{
modify_iri_style(t, prefixes, iri_style = c("short", "long", "mdlink", "html"))
}
\arguments{
\item{t}{Tibble whose IRIs are to be modified.}

\item{prefixes}{Tibble with "short" and "long" versions of the prefixes
for which the IRI style should be modified.}

\item{iri_style}{One of "short", "mdlink", "html" or "long".}
}
\value{
An copy of the input tibble \code{t} where the IRI style was modified.
}
\description{
Modify the style of IRIs in all columns of a tibble.
}
