% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{parse_construct_response}
\alias{parse_construct_response}
\title{Parse an HTTP response with content type "application/n-triples".}
\usage{
parse_construct_response(response)
}
\arguments{
\item{response}{SPARQL CONSTRUCT HTTP query response.}
}
\description{
Parse an HTTP response with content type "application/n-triples".
}
\keyword{internal}
