% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{parse_select_response}
\alias{parse_select_response}
\title{Parse an HTTP response with content type "application/sparql-results+json".}
\usage{
parse_select_response(response)
}
\arguments{
\item{response}{SPARQL SELECT HTTP query response.}
}
\description{
Parse an HTTP response with content type "application/sparql-results+json".
}
\keyword{internal}
