% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{query_result_to_tibble}
\alias{query_result_to_tibble}
\title{Convert a SPARQL SELECT query result into a tibble.}
\usage{
query_result_to_tibble(query_result, na_value = NA)
}
\arguments{
\item{query_result}{Object (list) to parse.}

\item{na_value}{Value with which to replace empty/missing fields.}
}
\description{
Converts the result of a parsed SPARQL SELECT query (a list object) into an
R tibble.
}
\keyword{internal}
