% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{rdf_term_to_string}
\alias{rdf_term_to_string}
\title{Convert an individual RDF term to an R string.}
\usage{
rdf_term_to_string(value, na_value = NA)
}
\arguments{
\item{value}{A list with \code{type} and \code{value} fields, or \code{NULL}.}

\item{na_value}{Value returned when \code{value} is NULL or an empty string.}
}
\description{
Convert an individual RDF term to an R string.
}
\keyword{internal}
