% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{sparql_describe}
\alias{sparql_describe}
\title{Run a SPARQL DESCRIBE query.}
\usage{
sparql_describe(
  endpoint,
  query,
  prefixes = NULL,
  request_method = c("POST", "GET"),
  http_extra_params = list(),
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{URL of SPARQL endpoint.}

\item{query}{SPARQL query as a string.}

\item{prefixes}{Optional data frame whose first two columns are
taken for short and long versions of base IRIs.}

\item{request_method}{HTTP method to use to submit the request.}

\item{http_extra_params}{Additional parameter/value pair to pass to the HTTP
request. E.g. some endpoints accept a "timeout"
argument, which could be passed via this argument.}

\item{verbose}{If \code{TRUE}, print query execution time.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A list with two tibbles: `edges` and `nodes`.
}\if{html}{\out{</div>}}
}
\description{
Executes a SPARQL DESCRIBE query and returns the results tibble with 3
columns: subject, predicate, object.

IRIs can be formatted in different styles using the \code{iri_style} argument.
}
\examples{
\dontshow{if (Sys.getenv("_R_SPARQLR_RUN_CONNECTED_EXAMPLES_") == 1) withAutoprint(\{ # examplesIf}

query <- "
# Retrieve all 'useful descriptions' associated with the 'cats' resource.
DESCRIBE <http://www.wikidata.org/entity/Q146>
"
endpoint <- "https://query.wikidata.org/sparql"

# Run the SPARQL query on the specified endpoint.
sparql_describe(endpoint, query)

# Same as above, but converting prefixes to their short form.
prefixes <- tibble::tibble(
  short = c("wd", "w3"),
  long = c(
    "http://www.wikidata.org/entity/",
    "http://www.w3.org/2000/01/rdf-schema#"
  )
)
sparql_describe(endpoint, query, prefixes = prefixes)
\dontshow{\}) # examplesIf}
}
