% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{sparql_query}
\alias{sparql_query}
\title{Run a SPARQL query}
\usage{
sparql_query(
  endpoint,
  query,
  prefixes = tibble::tribble(~short, ~long, "rdf",
    "http://www.w3.org/1999/02/22-rdf-syntax-ns#", "rdfs",
    "http://www.w3.org/2000/01/rdf-schema#"),
  http_params = list(),
  use_post = FALSE,
  add_prefixes = FALSE,
  iri_style = "short",
  na_value = NA,
  echo = FALSE
)
}
\arguments{
\item{endpoint}{URL of SPARQL endpoint.}

\item{query}{SPARQL query as a string.}

\item{prefixes}{Optional data frame whose first two columns are taken
for short and long versions of base IRIs.}

\item{http_params}{Additional parameter/value pair to pass with the HTTP
request.
Example: some endpoints accept a "timeout" argument.
This could be passed via this argument.}

\item{use_post}{Boolean to switch http protocol (default to GET). The
main benefit of POST is to allow for larger input query.}

\item{add_prefixes}{Boolean to add PREFIX declarations passed to the
\code{prefixes} argument to the SPARQL query.}

\item{iri_style}{One of "long", "short" , "html" or "mdlink" (markdown
link), to encode returned IRIs.}

\item{na_value}{Value with which to replace empty fields.}

\item{echo}{Boolean value to echo the SPARQL query before execution.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A tibble with the query results or NULL if the query
                returns nothing.
}\if{html}{\out{</div>}}
}
\description{
Run a SPARQL query, either SELECT, CONSTRUCT or DESCRIBE and
return the results as a tibble. Returned column names are the same as SPARQL
variables. Detection of column types relies on R built-in methods, not RDF
data types.

In the HTTP request, the "application/sparql-results+json" MIME type is
used, which is supported by most SPARQL endpoints.
}
\examples{
\dontrun{
 sparql_query(
   "https://query.wikidata.org/sparql",
   "PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
   SELECT ?message
   WHERE{
       wd:Q131303 rdfs:label ?message
       FILTER( LANG( ?message ) = 'en' )
   }"
 )
}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       SPARQL_ask() SPARQL_update()
}\if{html}{\out{</div>}}
}
