% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_sparql_prefix}
\alias{as_sparql_prefix}
\title{Convert a 2-column tibble into a multi-line SPARQL prefix string.}
\usage{
as_sparql_prefix(prefixes)
}
\arguments{
\item{prefixes}{A tibble with columns "short" and "long", containing
respectively the short and long forms of SPARQL prefixes.}
}
\description{
Convert a 2-column tibble into a multi-line SPARQL prefix string.
}
\keyword{internal}
