% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{as_sparql_prefix}
\alias{as_sparql_prefix}
\title{Converts a 2-column tibble into a SPARQL PREFIX string.}
\usage{
as_sparql_prefix(prefixes)
}
\arguments{
\item{prefixes}{2-column tibble containing the long and short form of
the SPARQL prefixes.}
}
\description{
Converts a 2-column tibble into a SPARQL PREFIX string.
}
\keyword{internal}
