% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{convert_iri_style}
\alias{convert_iri_style}
\title{Convert the style of IRIs from "long" or "short" into another form.}
\usage{
convert_iri_style(
  t,
  prefixes,
  iri_style = c("short", "long", "mdlink", "html"),
  replace_in_literal = FALSE
)
}
\arguments{
\item{t}{Tibble whose IRIs are to be modified.}

\item{prefixes}{Tibble with "short" and "long" versions of the
prefixes for which the IRI style should be
modified.}

\item{iri_style}{One of "short", "mdlink", "html" or "long".}

\item{replace_in_literal}{If TRUE, then replacement is also done within
literal strings, not just in IRIs.}
}
\value{
An copy of the input tibble \code{t} where the IRI style was modified.
}
\description{
Convert the style of IRIs in all columns of a tibble.
}
\examples{
# Create a sample tibble with long-form IRIs.
data <- tibble::tibble(
  composer = c(
    "<http://www.wikidata.org/entity/Q254>",
    "<http://www.wikidata.org/entity/Q255>",
    "<http://www.wikidata.org/entity/Q1268>"
  ),
  country = c(
    "<http://www.wikidata.org/entity/Q701614>",
    "<http://www.wikidata.org/entity/Q131964>",
    "<http://www.wikidata.org/entity/Q34266>"
  ),
  label = c(
    "Wolfgang Amadeus Mozart - http://example.org/Mozart",
    "Ludwig van Beethoven - http://example.org/Beethoven",
    "Frederic Chopin - http://example.org/Chopin)"
  )
)

# Define prefixes long-to-short correspondence.
prefixes <- tibble::tibble(
  short = c("wde", "ex"),
  long = c("http://www.wikidata.org/entity/", "http://example.org/")
)

# Convert IRIs to short form.
convert_iri_style(data, prefixes, iri_style = "short")

# Convert IRIs to markdown link format.
convert_iri_style(data, prefixes, iri_style = "mdlink")

# Convert IRIs to HTML link format.
convert_iri_style(data, prefixes, iri_style = "html")

# Also apply conversion to string literals.
convert_iri_style(
  data, prefixes,
  iri_style = "short",
  replace_in_literal = TRUE
)

# Apply different IRI styles to different columns.
convert_iri_style(
  data, prefixes,
  iri_style = c("short", "short", "mdlink"),
  replace_in_literal = TRUE
)

}
