% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_prefixes_from_file}
\alias{load_prefixes_from_file}
\title{Extract the list of PREFIXes from a SPARQL query file, and return it as
a tibble object.}
\usage{
load_prefixes_from_file(path)
}
\arguments{
\item{path}{path to SPARQL file.}
}
\description{
Extract the list of PREFIXes from a SPARQL query file, and return it as
a tibble object.
}
