% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{query_result_to_tibble}
\alias{query_result_to_tibble}
\title{Parse a \code{query_result} object - a nested list derived from the JSON returned
by a SPARQL query - and return it as an R tibble.}
\usage{
query_result_to_tibble(query_result, na_value = NA)
}
\arguments{
\item{query_result}{Object (list) to parse.}

\item{na_value}{Value with which to replace empty/missing fields.}
}
\description{
Parse a \code{query_result} object - a nested list derived from the JSON returned
by a SPARQL query - and return it as an R tibble.
}
\keyword{internal}
