% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{sparql_construct}
\alias{sparql_construct}
\title{Run a SPARQL CONSTRUCT query.}
\usage{
sparql_construct(
  endpoint,
  query,
  prefixes = NULL,
  request_method = c("POST", "GET"),
  http_extra_params = list(),
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{URL of SPARQL endpoint.}

\item{query}{SPARQL query as a string.}

\item{prefixes}{Optional data frame whose first two columns are
taken for short and long versions of base IRIs.}

\item{request_method}{HTTP method to use to submit the request.}

\item{http_extra_params}{Additional parameter/value pair to pass to the HTTP
request. E.g. some endpoints accept a "timeout"
argument, which could be passed via this argument.}

\item{verbose}{If \code{TRUE}, print query execution time.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A list with two tibbles: `edges` and `nodes`.
}\if{html}{\out{</div>}}
}
\description{
Executes a SPARQL CONSTRUCT query and returns the results as a list
containing two tibbles: one for edges (triples with IRIs) and one for nodes
with properties.
IRIs can be formatted in different styles using the \code{iri_style} argument.
}
\examples{
\dontshow{if (Sys.getenv("_R_SPARQLR_RUN_CONNECTED_EXAMPLES_") == 1) withAutoprint(\{ # examplesIf}

# Load a SPARQL construct query from an example file.
query <- system.file(
  "extdata", "example_construct.rq",
  package = "sparqlr"
) |> load_query_from_file()

# Run the SPARQL query on the specified endpoint.
sparql_construct(
  endpoint = "https://query.wikidata.org/sparql",
  query = query
)
\dontshow{\}) # examplesIf}
}
