% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{sparql_construct}
\alias{sparql_construct}
\title{Run a SPARQL CONSTRUCT query}
\usage{
sparql_construct(
  endpoint,
  query,
  prefixes = DEFAULT_PREFIXES,
  add_prefixes = FALSE,
  http_params = list(),
  use_post = FALSE,
  iri_style = "long",
  echo = FALSE
)
}
\arguments{
\item{endpoint}{URL of the SPARQL endpoint.}

\item{query}{SPARQL CONSTRUCT query as a string.}

\item{prefixes}{Optional data frame whose first two columns are short
and long versions of base IRIs.}

\item{add_prefixes}{Boolean to add PREFIX declarations from \code{prefixes} to
the query.}

\item{http_params}{Additional parameter/value pairs for the HTTP request.}

\item{use_post}{Boolean to use POST instead of GET for the HTTP request.}

\item{iri_style}{One of "long", "short", "html", or "mdlink" to encode
returned IRIs.}

\item{echo}{Print the SPARQL query that is executed in the terminal.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A list with two tibbles: `edges` and `nodes`.
}\if{html}{\out{</div>}}
}
\description{
Executes a SPARQL CONSTRUCT query and returns the results as a list
containing two tibbles: one for edges (triples with IRIs) and one for nodes
with properties.
IRIs can be formatted in different styles using the \code{iri_style} argument.
}
