% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\name{sparql_select}
\alias{sparql_select}
\title{Run a SPARQL SELECT query.}
\usage{
sparql_select(
  endpoint,
  query,
  prefixes = NULL,
  request_method = c("POST", "GET"),
  http_extra_params = list(),
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{URL of SPARQL endpoint.}

\item{query}{SPARQL query as a string.}

\item{prefixes}{Optional data frame whose first two columns are
taken for short and long versions of base IRIs.}

\item{request_method}{HTTP method to use to submit the request.}

\item{http_extra_params}{Additional parameter/value pair to pass to the HTTP
request. E.g. some endpoints accept a "timeout"
argument, which could be passed via this argument.}

\item{verbose}{If \code{TRUE}, print query execution time.}
}
\value{
A tibble with the query results or NULL if the query returns nothing.
}
\description{
Run a SPARQL query, either SELECT, CONSTRUCT or DESCRIBE and
return the results as a tibble. Returned column names are the same as SPARQL
variables. Detection of column types relies on R built-in methods, not RDF
data types.

In the HTTP request, the "application/sparql-results+json" MIME type is
used, which is supported by most SPARQL endpoints.
}
